+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to Merlin will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- A bug where assigning a worker all steps also assigned steps to the default worker
+
+### Added
+- Tests to make sure the default worker is being assigned properly
+
 ## [1.10.0]
 ### Fixed
 - Pip wheel wasn't including .sh files for merlin examples